+++ src/native/tabs.rs
@@ -165,7 +165,7 @@ where
     /// Sets the maximum height of the [`TabBar`](super::tab_bar::TabBar) of the
     /// [`Tabs`](Tabs).
     #[must_use]
-    pub fn tab_bar_max_height(mut self, max_height: u32) -> Self {
+    pub fn tab_bar_max_height(mut self, max_height: f32) -> Self {
         self.tab_bar = self.tab_bar.max_height(max_height);
         self
     }
@@ -173,7 +173,7 @@ where
     /// Sets the icon size of the [`TabLabel`](super::tab_bar::TabLabel) of the
     /// [`TabBar`](super::tab_bar::TabBar).
     #[must_use]
-    pub fn icon_size(mut self, icon_size: u16) -> Self {
+    pub fn icon_size(mut self, icon_size: f32) -> Self {
         self.tab_bar = self.tab_bar.icon_size(icon_size);
         self
     }
@@ -181,7 +181,7 @@ where
     /// Sets the text size of the [`TabLabel`](super::tab_bar::TabLabel) of the
     /// [`TabBar`](super::tab_bar::TabBar).
     #[must_use]
-    pub fn text_size(mut self, text_size: u16) -> Self {
+    pub fn text_size(mut self, text_size: f32) -> Self {
         self.tab_bar = self.tab_bar.text_size(text_size);
         self
     }
@@ -190,14 +190,14 @@ where
     /// [`TabLabel`](super::tab_bar::TabLabel) of the
     /// [`TabBar`](super::tab_bar::TabBar).
     #[must_use]
-    pub fn close_size(mut self, close_size: u16) -> Self {
+    pub fn close_size(mut self, close_size: f32) -> Self {
         self.tab_bar = self.tab_bar.close_size(close_size);
         self
     }
 
     /// Sets the padding of the tabs of the [`TabBar`](super::tab_bar::TabBar).
     #[must_use]
-    pub fn tab_label_padding(mut self, padding: u16) -> Self {
+    pub fn tab_label_padding(mut self, padding: f32) -> Self {
         self.tab_bar = self.tab_bar.padding(padding);
         self
     }
@@ -205,7 +205,7 @@ where
     /// Sets the spacing between the tabs of the
     /// [`TabBar`](super::tab_bar::TabBar).
     #[must_use]
-    pub fn tab_label_spacing(mut self, spacing: u16) -> Self {
+    pub fn tab_label_spacing(mut self, spacing: f32) -> Self {
         self.tab_bar = self.tab_bar.spacing(spacing);
         self
     }
